refactor: move @InformerConfig to more appropriate package
