+++ operator-framework/src/test/java/io/javaoperatorsdk/operator/dependent/dependentfilter/DependentFilterTestReconciler.java
@@ -2,9 +2,9 @@ package io.javaoperatorsdk.operator.dependent.dependentfilter;
 
 import java.util.concurrent.atomic.AtomicInteger;
 
+import io.javaoperatorsdk.operator.api.config.informer.InformerConfig;
 import io.javaoperatorsdk.operator.api.reconciler.*;
 import io.javaoperatorsdk.operator.api.reconciler.dependent.Dependent;
-import io.javaoperatorsdk.operator.processing.dependent.kubernetes.InformerConfig;
 
 @Workflow(dependents = {@Dependent(type = FilteredDependentConfigMap.class)})
 @ControllerConfiguration(informerConfig = @InformerConfig(onUpdateFilter = UpdateFilter.class))
